Add files via upload
Antakshari game (traditional Indian game ) .The game will be played between user and system. System has got intelligence in a way that 
it will tend to sing songs which user was previously unable to play provide it has that song in store. Code is running except you have to make a text file of songs. eg:
a----b
c----f
z----g